+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Updated
 
+- optimized _parse_all() which recognizes and parses any type of node source
 - parse expressionish elements don't have to start at beginning of line
 - lots of code cleanups